requirements.txt
python-socketio[client]
pysdl2
pysdl2-dll
